feat(web): render the device panel inline in Settings instead of navigating to /device
Extract the role/host-connection and local-app cards out of DeviceView into a
reusable DevicePanel and embed it in the Settings 'this device' tile, so it
stays inline like the other settings cards. DeviceView reuses the same panel.
